Conflicts:
	LabBook/src/main/java/sk/upjs/paz1c/gui/DeleteItemController.java
	LabBook/src/main/java/sk/upjs/paz1c/gui/DeleteNoteController.java
	LabBook/src/main/java/sk/upjs/paz1c/gui/DeleteProjectController.java
	LabBook/src/main/java/sk/upjs/paz1c/gui/DeleteTaskController.java
	LabBook/userData.xlsx

+++ LabBook/src/main/java/sk/upjs/paz1c/gui/DeleteItemController.java
@@ -4,6 +4,9 @@ import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
+import javafx.scene.input.KeyCode;
+import javafx.scene.input.KeyEvent;
+import sk.upjs.paz1c.business.UserIdentificationManager;
 import sk.upjs.paz1c.entities.Item;
 import sk.upjs.paz1c.entities.Project;
 import sk.upjs.paz1c.fxmodels.ItemFxModel;
@@ -30,13 +33,16 @@ public class DeleteItemController {
 	@FXML
 	void initialize() {
 
-		yesButton.setOnAction(new EventHandler<ActionEvent>() {
+		yesButton.setOnKeyPressed(new EventHandler<KeyEvent>() {
 
 			@Override
-			public void handle(ActionEvent event) {
-				itemDao.deleteItem(itemModel.getItem());
-				// projectDao.saveProject(projectModel.getProject());
-				yesButton.getScene().getWindow().hide();
+
+			public void handle(KeyEvent event) {
+				if (event.getCode().equals(KeyCode.ENTER)) {
+					itemDao.deleteItem(itemModel.getItem());
+					yesButton.getScene().getWindow().hide();
+				}
+
 			}
 		});
