+++ LabBook/src/main/java/sk/upjs/paz1c/gui/DeleteTaskController.java
@@ -4,6 +4,9 @@ import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
+import javafx.scene.input.KeyCode;
+import javafx.scene.input.KeyEvent;
+import sk.upjs.paz1c.business.UserIdentificationManager;
 import sk.upjs.paz1c.entities.Project;
 import sk.upjs.paz1c.entities.Task;
 import sk.upjs.paz1c.fxmodels.ProjectFxModel;
@@ -31,13 +34,15 @@ public class DeleteTaskController {
 	@FXML
 	void initialize() {
 
-		yesButton.setOnAction(new EventHandler<ActionEvent>() {
+		yesButton.setOnKeyPressed(new EventHandler<KeyEvent>() {
 
 			@Override
-			public void handle(ActionEvent event) {
-				taskDao.deleteTask(taskModel.getTask());
-				taskDao.saveTask(taskModel.getTask());
-				yesButton.getScene().getWindow().hide();
+			public void handle(KeyEvent event) {
+				if (event.getCode().equals(KeyCode.ENTER)) {
+					taskDao.deleteTask(taskModel.getTask());
+					taskDao.saveTask(taskModel.getTask());
+					yesButton.getScene().getWindow().hide();
+				}
 			}
 		});
